requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
codefind==0.1.3
discord-py-interactions==5.11.0
discord-typings==0.8.0
emoji==2.11.0
frozenlist==1.4.1
idna==3.6
jurigged==0.5.7
multidict==6.0.5
ovld==0.3.2
sentry-sdk==1.44.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
yarl==1.9.4
